Remove timestamp file
