+++ lib/constants.js
@@ -65,11 +65,18 @@ const STATS = {
  */
 const IMG_SCORE_THRESHOLD = 20000;
 
+/**
+ * Example: Mar 26 10:56AM
+ */
+const SIMPLE_DATE_FORMAT = 'MMM d h:mma';
+
+
 module.exports = {
     RESPONSES,
     VALID_MAPS,
     PUBG_APP_ID,
     ALLOWED_EXT,
     STATS,
-    IMG_SCORE_THRESHOLD
+    IMG_SCORE_THRESHOLD,
+    SIMPLE_DATE_FORMAT
 }
