+++ lib/discord.js
@@ -1,10 +1,9 @@
 const Discord = require("discord.js");
 const logger = require("./log")('discord');
 const app = require("../app");
-const {deleteFile} = require('./utils');
-const { getTimestamp, SIMPLE_DATE_FORMAT } = require('./timestamp');
+const {getTimestamp} = require('./utils');
 const db = require('./db_utils');
-const {RESPONSES} = require('./constants');
+const {RESPONSES, SIMPLE_DATE_FORMAT} = require('./constants');
 
 
 /**
