+++ lib/utils.js
@@ -4,8 +4,9 @@ const path = require('path');
 const mime = require('mime-types');
 const axios = require('axios');
 const jimp = require('jimp');
-
+const { format, parseISO } = require('date-fns');
 const {ALLOWED_EXT} = require('./constants');
+
 const logger = require('./log.js')('utils');
 
 /**
@@ -112,6 +113,21 @@ const getImage = async(url) => {
 }
 
 
+ 
+ /**
+  * Generates timestamp in given format
+  * @param {*} dateFormat 
+  * @param {*} inputDate 
+  * @returns current time OR input date in given format
+  */
+const getTimestamp = (dateFormat, inputDate) => {
+	let date = new Date();
+	if(inputDate){
+		date = parseISO(inputDate)
+	}
+	return format(date, dateFormat);
+}
+
 
 module.exports = {
 	deleteFile,
@@ -119,6 +135,7 @@ module.exports = {
 	hashFile,
 	downloadFile,
 	getImage,
-	hashText
+	hashText,
+	getTimestamp
 }
